improve quest testing service

+++ lib/ui/views/active_quest_standalone_ui/active_distance_estimate_quest/active_distance_estimate_quest_viewmodel.dart
@@ -33,7 +33,7 @@ class ActiveDistanceEstimateQuestViewModel extends ActiveQuestBaseViewModel {
 
   double? _currentSpeed;
   double? get currentSpeed => _currentSpeed;
-  double? get currentAccuracy => _geolocationService.currentGPSAccuracy;
+  int? get currentAccuracy => _geolocationService.currentGPSAccuracy;
 
   bool startedQuest = false;
   bool questSuccessfullyFinished = false;
